overhauled file structure
