+++ chapter/testing/test_3.py
@@ -1,16 +1,15 @@
-from unittest.mock import patch
 from flask import url_for
 from flask_testing import TestCase
-from application import application
+import requests_mock
+from app import app
 
 class TestBase(TestCase):
-
     def create_app(self):
         return app
 
 class TestResponse(TestBase):
 
-        def test_striker_position(self):
+        def chapter_test1(self):
             with patch('random.randrange') as r:
                 r.return_value = 0
                 response = self.client.get(url_for('role'))
@@ -18,7 +17,7 @@ class TestResponse(TestBase):
 
 class TestResponse(TestBase):
 
-        def test_striker_position(self):
+        def chapter_test2(self):
             with patch('random.randrange') as r:
                 r.return_value = 1
                 response = self.client.get(url_for('role'))
@@ -26,7 +25,7 @@ class TestResponse(TestBase):
 
 class TestResponse(TestBase):
 
-        def test_striker_position(self):
+        def chapter_test3(self):
             with patch('random.randrange') as r:
                 r.return_value = 2
                 response = self.client.get(url_for('role'))
@@ -34,7 +33,7 @@ class TestResponse(TestBase):
 
 class TestResponse(TestBase):
 
-        def test_striker_position(self):
+        def chapter_test4(self):
             with patch('random.randrange') as r:
                 r.return_value = 3
                 response = self.client.get(url_for('role'))
