--- frontend/application/__init__.py
@@ -1,13 +1,0 @@
-from flask import Flask
-from flask_sqlalchemy import SQLAlchemy
-from os import getenv
-
-app = Flask(__name__)
-
-
-app.config['SQLALCHEMY_DATABASE_URI'] = getenv("DATABASE_URI")
-app.config['SECRET_KEY'] = getenv('SECRET_KEY')
-
-db = SQAlchemy(app)
-
-from application import routes
