+++ frontend/testing/test_1.py
@@ -1,7 +1,6 @@
-from unittest.mock import patch
 from flask import url_for
 from flask_testing import TestCase
-
+import requests_mock
 from app import app
 
 class TestBase(TestCase):
